+++ CHANGELOG.md
@@ -64,6 +64,7 @@ The format is based on [Keep a Changelog].
 
 ### Changed
 
+-   `pulse.samplers` module has now been moved to `pulse.pulse_lib.samplers`.
 -   The number of memory slots required will now be inferred from the supplied
     schedules if `memory_slots` is not supplied.
 -   All circuit drawers now express most commonly used fractions